backend/requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.7.0
appnope==0.1.3
asgiref==3.6.0
astroid==2.9.2
asttokens==2.1.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==21.4.0
Automat==22.10.0
backcall==0.2.0
backoff==1.10.0
backports.entry-points-selectable==1.1.1
bandit==1.7.1
beautifulsoup4==4.11.1
black==23.1.0
boto==2.49.0
CacheControl==0.12.10
cachy==0.3.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cql==1.4.0
crashtest==0.3.1
cryptography==38.0.3
cssselect==1.2.0
Cython==0.29.32
datadog==0.43.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.3.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
Faker==13.3.0
fastapi==0.97.0
filelock==3.4.0
flake8==4.0.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.24
googleapis-common-protos==1.56.1
greenlet==1.1.2
grpcio==1.46.1
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.11.2
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipykernel==6.17.0
ipython==8.6.0
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==7.4.4
jupyter_core==4.11.2
keyring==21.8.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
motor==3.1.2
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==1.0.0
neovim==0.3.1
nest-asyncio==1.5.6
numpy==1.23.1
opentelemetry-api==1.9.0
opentelemetry-distro==0.28b0
opentelemetry-exporter-otlp==1.9.0
opentelemetry-exporter-otlp-proto-grpc==1.9.0
opentelemetry-exporter-otlp-proto-http==1.9.0
opentelemetry-instrumentation==0.28b0
opentelemetry-proto==1.9.0
opentelemetry-sdk==1.9.0
opentelemetry-semantic-conventions==0.28b0
packaging==23.0
pandas==1.5.1
parsel==1.7.0
parso==0.8.3
passlib==1.7.4
pastel==0.2.1
pathspec==0.11.0
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.4.1
pluggy==1.0.0
poetry==1.1.12
poetry-core==1.0.7
prompt-toolkit==3.0.32
Protego==0.2.1
protobuf==3.20.1
psutil==5.9.3
psycopg2==2.9.3
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.9
PyDispatcher==2.0.6
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.3.0
pylev==1.4.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pymongo==4.3.3
pynvim==0.4.3
pyOpenSSL==22.1.0
pyparsing==3.0.6
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2020.1
PyYAML==6.0
pyzmq==24.0.1
queuelib==1.6.2
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
Scrapy==2.7.1
service-identity==21.1.0
shellingham==1.4.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
sqlparse==0.3.1
stack-data==0.6.0
starlette==0.27.0
stevedore==3.5.0
thrift==0.16.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.7.2
tornado==6.2
traitlets==5.5.0
Twisted==22.10.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.22.0
validate-docbr==1.10.0
virtualenv==20.10.0
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
zope.interface==5.5.1
